% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regrans_ancova.R
\name{regrans_ancova}
\alias{regrans_ancova}
\title{Implements an ANCOVA for Parameters Comparisons of Segmented
    Linear Regressions Based on a \code{regrans} Object}
\usage{
regrans_ancova(obj)
}
\arguments{
\item{obj}{\code{regrans} object, result of \code{regrans} function.}

\item{...}{further arguments passed to or from other methods to
plot.}
}
\description{
Parameter comparisons of a segmented linear regression
    model based on an interactive \code{regrans} routine.
}
\author{
Rodrigo Sant'Ana \email{rsantana@univali.br}, Fernando
    Mayer \email{fernando.mayer@ufpr.br}
}
