% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regranslm.R
\name{regranslm}
\alias{regranslm}
\title{Fits Segmented Linear Regression Models}
\usage{
regranslm(x, y, n.min = 5)
}
\arguments{
\item{x}{a numeric vector.}

\item{y}{a numeric vector.}

\item{n.min}{minimum number of points (x,y) to be considered in a
linear model.}
}
\value{
a data frame containing the results of each model fitted.
}
\description{
Fit segmented linear models
}
\author{
Rodrigo Sant'Ana \email{rsantana@univali.br}, Fernando
    Mayer \email{fernando.mayer@ufpr.br}
}
