% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.regrans.R
\name{summary.regrans}
\alias{summary.regrans}
\title{Summarizing Segmented Linear Regression Fits}
\usage{
\method{summary}{regrans}(object, correlation = FALSE, symbolic.cor = FALSE,
  ...)
}
\arguments{
\item{object}{an object of class \code{regrans}, usually, a result
of a call to \code{regrans} function.}

\item{correlation}{logical; if ‘TRUE’, the correlation matrix of the
estimated parameters is returned and printed.}

\item{symbolic.cor}{logical. If ‘TRUE’, print the correlations in a
symbolic form (see ‘symnum’) rather than as numbers.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
The function ‘summary.regrans’ computes and returns a list
    of summaries statistics of both (left and right lm) fitted
    segmented regression given in ‘object’. For more details see
    also the help of summary.lm.
}
\description{
"summary" method for class \code{regrans}.
}
\author{
Rodrigo Sant'Ana \email{rsantana@univali.br}, Fernando
    Mayer \email{fernando.mayer@ufpr.br}
}

