% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Regrans_package.R
\docType{data}
\name{simdata}
\alias{simdata}
\title{Simulated data}
\format{a data frame with 50 rows and 2 columns:
\itemize{
  \item x: explanatory variable.
  \item y: response variable.
}}
\description{
A simple simulated data just for a good example.
}
\keyword{datasets}
